builbot: master: Store build binaries on local server instead of master
Uploading to master and then downloading from master takes time
(download is ~75 seconds) and makes no sense if build machine and boot
server are in the same local network.  Store the build artifacts for
booting in build machine public_html and download them with wget over
HTTP.  Compression of vmlinux and other files is also not really needed,
because there is plenty of space on the build server.

This nicely speeds things up:
1. Build:
 - before: 15 s (xz) + 25 s (upload) = 40 s
 - after: 0 s (yay!)

2. Download/boot:
 - before: 75 s
 - after: 4 s

Signed-off-by: Krzysztof Kozlowski <krzk@kernel.org>